You and I both know I'm not gonna document all these changes right now. I'll make some docs later...with flow charts..and pretty colors

+++ core/Counter.h
@@ -4,9 +4,9 @@
 class Counter
 {
 public:
-	double GetDelta() const
+	float GetDelta() const
 	{
-		return delta_;
+		return static_cast<float>(delta_);
 	}
 
 	void Update()
